chore(test): share article-completion recorder and dispose barriers
